requirements.txt
dnspython
requests
pyfiglet
